apps/ml-ocr/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic==2.*
python-multipart
pillow
pytesseract
pdfminer.six
pymupdf
numpy
